+++ README.md
@@ -1,49 +1,1 @@
-# Project Title
-
-A brief one or two sentence description of the project.
-
-## Table of Contents
-
-- [Project Title](#project-title)
-  - [Table of Contents](#table-of-contents)
-  - [Installation](#installation)
-  - [Usage](#usage)
-  - [Contributing](#contributing)
-  - [Credits](#credits)
-  - [License](#license)
-
-## Installation
-
-Step-by-step instructions on how to install and set up the project locally. This may include:
-
-- Prerequisites (e.g., Node.js version, package manager)
-- Cloning the repository
-- Installing dependencies
-- Setting up environment variables
-- Running a development server
-
-## Usage
-
-Instructions on how to use the project, including:
-
-- Main features and functionality
-- Examples or code snippets
-- Configuration options
-- Deployment instructions
-
-## Contributing
-
-Guidelines for contributing to the project, such as:
-
-- Reporting issues
-- Opening pull requests
-- Code style and conventions
-- Testing instructions
-
-## Credits
-
-List of contributors, resources, libraries, or assets used in the project.
-
-## License
-
-Information about the license under which the project is distributed.
+# replace this
