+++ frontend/.gitattributes
@@ -6,7 +6,8 @@
 /.github/pull_request_template.md linguist-generated
 /.github/workflows/build.yml linguist-generated
 /.github/workflows/pull-request-lint.yml linguist-generated
-/.github/workflows/upgrade.yml linguist-generated
+/.github/workflows/release.yml linguist-generated
+/.github/workflows/upgrade-main.yml linguist-generated
 /.gitignore linguist-generated
 /.mergify.yml linguist-generated
 /.npmignore linguist-generated
