+++ frontend/.github/pull_request_template.md
@@ -1,0 +1,1 @@
+Fixes #
