+++ frontend/.gitignore
@@ -34,13 +34,16 @@ jspm_packages/
 junit.xml
 /coverage/
 !/.github/workflows/build.yml
+/dist/changelog.md
+/dist/version.txt
+!/.github/workflows/release.yml
 !/.mergify.yml
-!/.github/workflows/upgrade.yml
+!/.github/workflows/upgrade-main.yml
 !/.github/pull_request_template.md
 !/.prettierignore
 !/.prettierrc.json
 !/.npmrc
-!/test/
+!/tests/
 !/tsconfig.json
 !/tsconfig.dev.json
 !/src/
