+++ frontend/.mergify.yml
@@ -1,0 +1,24 @@
+# ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
+
+queue_rules:
+  - name: default
+    update_method: merge
+    conditions:
+      - "#approved-reviews-by>=1"
+      - -label~=(do-not-merge)
+      - status-success=build
+pull_request_rules:
+  - name: Automatic merge on approval and successful build
+    actions:
+      delete_head_branch: {}
+      queue:
+        method: squash
+        name: default
+        commit_message_template: |-
+          {{ title }} (#{{ number }})
+
+          {{ body }}
+    conditions:
+      - "#approved-reviews-by>=1"
+      - -label~=(do-not-merge)
+      - status-success=build
