+++ frontend/.npmignore
@@ -4,10 +4,12 @@
 junit.xml
 /coverage/
 permissions-backup.acl
+/dist/changelog.md
+/dist/version.txt
 /.mergify.yml
 /.prettierignore
 /.prettierrc.json
-/test/
+/tests/
 /tsconfig.dev.json
 /src/
 !/lib/
