+++ frontend/.npmrc
@@ -1,0 +1,3 @@
+# ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
+
+resolution-mode=highest
