+++ frontend/.prettierignore
@@ -1,0 +1,1 @@
+# ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
