+++ frontend/.prettierrc.json
@@ -1,0 +1,4 @@
+{
+  "singleQuote": true,
+  "overrides": []
+}
