+++ frontend/.projen/files.json
@@ -5,7 +5,8 @@
     ".github/pull_request_template.md",
     ".github/workflows/build.yml",
     ".github/workflows/pull-request-lint.yml",
-    ".github/workflows/upgrade.yml",
+    ".github/workflows/release.yml",
+    ".github/workflows/upgrade-main.yml",
     ".gitignore",
     ".mergify.yml",
     ".npmignore",
