+++ frontend/package.json
@@ -2,7 +2,8 @@
   "name": "name",
   "description": "A cool typescript project",
   "scripts": {
-    "build": "npx projen build",
+    "build": "next build",
+    "bump": "npx projen bump",
     "clobber": "npx projen clobber",
     "compile": "npx projen compile",
     "default": "npx projen default",
@@ -12,39 +13,55 @@
     "post-compile": "npx projen post-compile",
     "post-upgrade": "npx projen post-upgrade",
     "pre-compile": "npx projen pre-compile",
+    "release": "npx projen release",
     "test": "npx projen test",
     "test:watch": "npx projen test:watch",
+    "unbump": "npx projen unbump",
     "upgrade": "npx projen upgrade",
     "watch": "npx projen watch",
-    "projen": "npx projen"
+    "projen": "npx projen",
+    "preinstall": "npx only-allow pnpm",
+    "lint": "eslint . --fix --max-warnings 0",
+    "dev": "next dev",
+    "start": "next start"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.10",
+    "@types/jest": "^29.5.12",
     "@types/node": "^18",
+    "@types/react": "^18",
     "@typescript-eslint/eslint-plugin": "^6",
     "@typescript-eslint/parser": "^6",
+    "autoprefixer": "^10.4",
     "constructs": "^10.0.0",
     "eslint": "^8",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.7.0",
     "jest-junit": "^15",
-    "nx": "^18.2",
-    "prettier": "^3.2",
+    "postcss": "^8.4",
+    "prettier": "^3.2.5",
     "projen": "0.80.19",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.3.2"
+    "standard-version": "^9",
+    "tailwindcss": "^3.4",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.4"
+  },
+  "dependencies": {
+    "next": "^14.1",
+    "react": "^18",
+    "react-dom": "^18"
   },
   "pnpm": {},
+  "main": "lib/index.js",
   "license": "MIT",
   "version": "0.0.0",
   "jest": {
     "testMatch": [
       "<rootDir>/src/**/__tests__/**/*.ts?(x)",
-      "<rootDir>/(test|src)/**/*(*.)@(spec|test).ts?(x)"
+      "<rootDir>/(tests|src)/**/*(*.)@(spec|test).ts?(x)"
     ],
     "clearMocks": true,
     "collectCoverage": true,
@@ -83,5 +100,6 @@
       ]
     }
   },
+  "types": "lib/index.d.ts",
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
 }
