+++ frontend/tests/hello.test.ts
@@ -1,0 +1,5 @@
+import { Hello } from '../src';
+
+test('hello', () => {
+  expect(new Hello().sayHello()).toBe('hello, world!');
+});
