+++ frontend/tsconfig.dev.json
@@ -1,8 +1,6 @@
 // ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 {
   "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "lib",
     "alwaysStrict": true,
     "declaration": true,
     "esModuleInterop": true,
@@ -10,7 +8,9 @@
     "inlineSourceMap": true,
     "inlineSources": true,
     "lib": [
-      "es2019"
+      "dom",
+      "dom.iterable",
+      "esnext"
     ],
     "module": "CommonJS",
     "noEmitOnError": false,
@@ -25,10 +25,33 @@
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
     "stripInternal": true,
-    "target": "ES2019"
+    "target": "ES2019",
+    "rootDir": ".",
+    "allowJs": true,
+    "skipLibCheck": true,
+    "noEmit": true,
+    "moduleResolution": "bundler",
+    "isolatedModules": true,
+    "jsx": "preserve",
+    "incremental": true,
+    "paths": {
+      "@/*": [
+        "./src/*"
+      ]
+    }
   },
   "include": [
-    "src/**/*.ts"
+    "src/**/*.ts",
+    "tests/**/*.ts",
+    "next-env.d.ts",
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    ".next/types/**/*.ts",
+    ".projenrc.ts",
+    "projenrc/**/*.ts"
   ],
-  "exclude": []
+  "exclude": [
+    "node_modules",
+    "node_modules"
+  ]
 }
