+++ src/index.ts
@@ -1,0 +1,5 @@
+export class Hello {
+  public sayHello() {
+    return 'hello, world!';
+  }
+}
