--- tests/dummy.test.ts
@@ -1,4 +1,0 @@
-// assert true == true to pass one test
-test('dummy test', () => {
-  expect(true).toStrictEqual(true);
-});
