+++ tsconfig.dev.json
@@ -8,9 +8,7 @@
     "inlineSourceMap": true,
     "inlineSources": true,
     "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
+      "es2019"
     ],
     "module": "CommonJS",
     "noEmitOnError": false,
@@ -25,33 +23,15 @@
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
     "stripInternal": true,
-    "target": "ES2019",
-    "rootDir": ".",
-    "allowJs": true,
-    "skipLibCheck": true,
-    "noEmit": true,
-    "moduleResolution": "bundler",
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "incremental": true,
-    "paths": {
-      "@/*": [
-        "./src/*"
-      ]
-    }
+    "target": "ES2019"
   },
   "include": [
     "src/**/*.ts",
-    "tests/**/*.ts",
-    "next-env.d.ts",
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    ".next/types/**/*.ts",
+    "test/**/*.ts",
     ".projenrc.ts",
     "projenrc/**/*.ts"
   ],
   "exclude": [
-    "node_modules",
     "node_modules"
   ]
 }
